Added SearchEvaluator to QueryEvaluator class
A new evaluator, `SearchEvaluator`, has been added to the `QueryEvaluator` class. This evaluator is designed to handle search queries. A new file, `SearchEvaluator.cs`, has been created to house the `SearchEvaluator` class, which implements the `IEvaluator` interface. The `SearchEvaluator` class is a singleton, with a private constructor and a public static `Instance` property. The `GetQuery` method in the `SearchEvaluator` class takes an `IQueryable<T>` and an `IQuery<T>` as parameters, and returns an `IQueryable<T>`. This method uses the `Search` extension method on the `IQueryable<T>` object, passing in the `SearchInfoList` from the `SearchBuilder` of the `IQuery<T>` object. The `SearchEvaluator` class uses the `Craft.QuerySpec.Contracts` and `Craft.QuerySpec.Extensions` namespaces.