Refactor SearchInfo and SearchInfoJsonConverter classes
Refactored the `SearchInfo` and `SearchInfoJsonConverter<T>` classes for improved readability and reliability. Added the `System.Reflection` namespace to `SearchInfo.cs` for retrieving metadata information. Replaced `ArgumentNullException.ThrowIfNull` and `ArgumentException.ThrowIfNullOrEmpty` methods with explicit null or empty checks for `searchItem` and `searchTerm` in the `SearchInfo` constructor. Reformatted the empty constructor `SearchInfo()` to be on a single line for conciseness. In the `Write` method of `SearchInfoJsonConverter<T>`, changed the way the member name of `SearchItem` is retrieved to use the `GetPropertyInfo().Name` method for increased reliability.