Add QuerySelectExtensionsTests and MyResult class
Added a new test class `QuerySelectExtensionsTests` to the `Craft.QuerySpec.Tests.Core` namespace. This class contains unit tests for the `Select` method of the `Query` class, testing it under various scenarios. The tests follow the Arrange-Act-Assert pattern and use the `FluentAssertions` library for assertions. Also added a new model class `MyResult` with properties `Id`, `Name`, and `Age`, which is used in some of the tests.