Add SearchExtensionTests class for Search method
Added a new test class `SearchExtensionTests` in the `Craft.QuerySpec.Tests.Extensions` namespace to test the `Search` extension method for the `Company` model. The class is part of the `SystemTestCollectionDefinition` collection and uses an in-memory database for testing. Three test methods are included to test different search criteria scenarios. All tests use the FluentAssertions library for assertions.