Replace `DisableAuditingAttribute` with `NoAuditAttribute`
The `DisableAuditingAttribute` class has been removed and replaced with the `NoAuditAttribute` class. The new class serves the same purpose of disabling auditing for classes it's applied to, but does not contain a method to check if a class is derived from a certain base class. The `HasAuditAttributeDisabled` method in the `TypeExtensions` class has been updated to check for the `NoAuditAttribute` instead. The `DisableAuditingAttributeTests` class has been renamed to `NoAuditAttributeTests` and all instances of `DisableAuditingAttribute` in this class have been replaced with `NoAuditAttribute`. The same replacement has been done in the `TypeExtensionsTests` class.