Renamed `HasAuditAttributeDisabled` to `HasNoAuditAttribute`
Renamed the method `HasAuditAttributeDisabled` in `TypeExtensions.cs` to `HasNoAuditAttribute` and updated all references to this method in `AuditingHelpers.cs`, `DisableAuditingAttributeTests.cs`, and `TypeExtensionsTests.cs`. The functionality remains the same, it still checks if the `NoAuditAttribute` is defined for a given type. In `DisableAuditingAttributeTests.cs`, the method `HasDisableAuditingAttribute` has been removed and replaced with the extension method `HasNoAuditAttribute`. The test method names and assertions have also been updated to reflect this change.

Renamed `HasAuditAttributeDisabled` to `HasNoAuditAttribute`

The method `HasAuditAttributeDisabled` in `TypeExtensions.cs` has been renamed to `HasNoAuditAttribute`. All references to this method in `AuditingHelpers.cs`, `DisableAuditingAttributeTests.cs`, and `TypeExtensionsTests.cs` have been updated accordingly. In `DisableAuditingAttributeTests.cs`, the method `HasDisableAuditingAttribute` has been removed and replaced with direct calls to the `HasNoAuditAttribute` extension method on the `Type` instances. Test method names in `DisableAuditingAttributeTests.cs` have also been updated to reflect the new method name.