Added new configuration test files and set copy rules
Added two new files, `ConfigurationStoreTestSettings.json` and `ConfigurationStoreTestSettings_NoDefaults.json`, to the project. These files are configured to be copied to the output directory if they are newer than the existing files in the directory. This is achieved by adding a new `ItemGroup` to the project file and setting the `CopyToOutputDirectory` property for these files as `PreserveNewest`.