Add CreateUriWithReturnPathToThisPage method
Introduce a new method in NavigationExtensions to construct URLs
with a return path to the current page. Update CreditSummary,
LatestJD, and LatestResume to use this method for URL
construction, enhancing code reuse and clarity. Also,
add a using directive for Domain.Financials.Credits in
BuyCredits.razor while maintaining route attributes.